src: Rename some dirs and files
