+++ src/App.css
@@ -1,7 +1,7 @@
 .main-menu {
     height: 100vh;
     background: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)),
-                url('background.png') no-repeat center center;
+                url('images/background.png') no-repeat center center;
     background-size: cover;
     color: #fff;
     text-align: center;
