+++ src/components/Option.js
@@ -1,0 +1,10 @@
+import { useState } from "react";
+
+
+export default function StoryMode() {
+	return (
+		<h1>
+			THIS IS OPTIONS PAGE
+		</h1>
+	)
+}
