+++ src/components/StoryMode.js
@@ -1,0 +1,10 @@
+import { useState } from "react";
+
+
+export default function StoryMode() {
+	return (
+		<h1>
+			THIS IS STORYMODE PAGE
+		</h1>
+	)
+}
